+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+
+- Changed how chapters are retrieved on Fanfox source. Fanfox is slowly removing the auxiliar javascript file with the chapter list and that change was breaking Foxy.
 
 ## [0.7.0] - 2018-05-12
 ### Added